Renamed serde feature and expanded prelude::* imports
As the `serde` feature name collides with the `serde` external
  dependency, the `bitcoin` package could not imply other feature
logistics. By renaming it to something else (`enable-serde`), mutatis
mutandis, this "problem" no longer applies and serde doesn't have to be
renamed internally (which was called `actual_serde`).

Without this change I was getting weird errors from rust-analyzer on
vscode. No idea why, but this has solved it.

Note: I also removed the `macro_use` from the `lib.rs` because I personally
prefer to import the derive macros manually.

I personally prefer to deal with symbols, specially if those symbols
vary according to the architecture (or feature flags), in an "expanded"
way, individually referring or importing them to the file.

In `network/message_blockdata.rs` I also renamed a match ref symbol,
since that name would be somewhat ambiguous with the pre-existing `w`
symbol (from the method parameter) on the call-site. This was
interesting because if I tried expanding the macro with ra and replacing the
result, I would get an error since it would shadow the previous `w`. But
the macro is able to work around this by knowing that the `w`, literally
stated inside the macro, actually refers to the shadowed `w` symbol.
Anyway I just renamed the symbol and prevented the shadowing.

None of those changes were tested and no documentation were updated.
I intend to deviate a lot from upstream so the docs will be very off for
a long time.